form elems refacto
