+++ app.territoiresentransitions.react/src/app/pages/collectivite/PlanActions/defaultCategorie.ts
@@ -1,0 +1,9 @@
+import {FicheActionCategorieStorable} from 'storables/FicheActionCategorieStorable';
+
+export const defaultCategorie = new FicheActionCategorieStorable({
+  uid: '',
+  epci_id: '',
+  nom: 'sans categorie',
+  parent_uid: '',
+  fiche_actions_uids: [],
+});
